src/auth/requirements.txt
astroid==2.15.4
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Flask==2.3.2
google-api-core==1.34.0
google-auth==2.17.3
google-cloud-pubsub==2.1.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.48.2
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
libcst==0.4.9
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
parso==0.8.3
platformdirs==3.5.0
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pylint==2.17.4
PyMySQL==1.0.3
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
wrapt==1.15.0
zipp==3.15.0
